Extract getDefaultMiddleware to its own file
