+++ src/index.ts
@@ -7,3 +7,4 @@ export * from './createAction'
 export * from './createReducer'
 export * from './createSlice'
 export * from './serializableStateInvariantMiddleware'
+export * from './getDefaultMiddleware'
